refactor: Update linux_enum.sh to list all users' home directories
This commit modifies the linux_enum.sh script to list all users' home directories instead of just the current user's home directory. This change is made by replacing the command that retrieves the current user's home directory with a command that retrieves all users from the /etc/passwd file.

The previous implementation only listed the current user's home directory, which was stored in the $user_has_home variable. However, this variable was not always set correctly, leading to incorrect results. By using the command "cut -d: -f1 /etc/passwd", we can retrieve a list of all users and their corresponding home directories.

This change improves the functionality of the script by providing a more comprehensive view of the home directories on the system.